Fix built form page not found delete compose file
The previously built container was not copying the form page built and
therefore was throwing a 404

Also remove the docker-compose file as it is misleading. The production
build refers to the Dockerfile and the docker-compose here that is used for
development causes misleading in configs. Use
sourcify/environments/repository.yaml instead, removing image: and
referring the build: field to this directory.