Don't mount persistent volume, remove docker-cmpse
Since the Dockerfile was mounting a volume /h5ai in the last lines, the
chagnes made in the container does not take effect in the staging. I.e.
we unzip and copy things to /h5ai in the container but in the end the
mount /h5ai mounts this folder from the host machine version, which is
outdated. Because of this the changes made in the Dockerfile and the
repo is not taking effect on the staging.

Also remove the docker-compose.yaml (wasn't removed on prev. commit as
intended). Remove .zip as this is created build time with the
containers.